# web-api/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Webcomics Platform API
  version: 0.1.0
  description: |
    HTTP API for the webcomics generation platform. Two route groups:
      * `/users/*`  — user CRUD (boilerplate carry-over)
      * `/api/*`    — pipeline runs, templates, assets
servers:
  - url: http://localhost:8080
paths:
...
  /api/formats:
    get:
      operationId: listFormats
      summary: List available visual+composition format presets.
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FormatView' }
        default: { $ref: '#/components/responses/Error' }
  /api/projects:
    get:
      operationId: listProjects
      summary: List projects
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProjectView' }
        default: { $ref: '#/components/responses/Error' }
    post:
      operationId: createProject
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectBody' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IDResponse' }
        default: { $ref: '#/components/responses/Error' }
  /api/projects/{id}:
    get:
      operationId: getProject
      summary: Get a project detail (incl. characters/environments/plot)
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectDetailView' }
        default: { $ref: '#/components/responses/Error' }
    put:
      operationId: updateProject
      summary: Update project name/description/archived
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectBody' }
      responses:
        '204':
          description: no content
        default: { $ref: '#/components/responses/Error' }
    delete:
      operationId: deleteProject
      summary: Delete a project (cascades characters/environments/plot)
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '204':
          description: no content
        default: { $ref: '#/components/responses/Error' }
  /api/projects/{id}/characters:
    post:
      operationId: createCharacter
      summary: Create a character in this project
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CharacterBody' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IDResponse' }
        default: { $ref: '#/components/responses/Error' }
  /api/characters/{id}:
    put:
      operationId: updateCharacter
      summary: Update a character
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CharacterBody' }
      responses:
        '204':
          description: no content
        default: { $ref: '#/components/responses/Error' }
    delete:
      operationId: deleteCharacter
      summary: Delete a character
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '204':
          description: no content
        default: { $ref: '#/components/responses/Error' }
  /api/projects/{id}/environments:
    post:
      operationId: createEnvironment
      summary: Create an environment in this project
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnvironmentBody' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IDResponse' }
        default: { $ref: '#/components/responses/Error' }
  /api/environments/{id}:
    put:
      operationId: updateEnvironment
      summary: Update an environment
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnvironmentBody' }
      responses:
        '204':
          description: no content
        default: { $ref: '#/components/responses/Error' }
    delete:
      operationId: deleteEnvironment
      summary: Delete an environment
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '204':
          description: no content
        default: { $ref: '#/components/responses/Error' }
  /api/projects/{id}/social-accounts:
    post:
      operationId: createSocialAccount
      summary: Create a social account in this project
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SocialAccountBody' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IDResponse' }
        default: { $ref: '#/components/responses/Error' }
  /api/social-accounts/{id}:
    put:
      operationId: updateSocialAccount
      summary: Update a social account
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SocialAccountBody' }
      responses:
        '204':
          description: no content
        default: { $ref: '#/components/responses/Error' }
    delete:
      operationId: deleteSocialAccount
      summary: Delete a social account
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '204':
          description: no content
        default: { $ref: '#/components/responses/Error' }
  /api/projects/{id}/plot:
    put:
      operationId: upsertPlot
      summary: Create or update the project's plot
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlotBody' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IDResponse' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  parameters:
    PathID:
      in: path
      name: id
      required: true
      schema: { type: string }
  responses:
    Error:
      description: error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    IDResponse:
      type: object
      required: [id]
      properties:
        id: { type: string }
...
    FormatView:
      type: object
      required: [id, name, scope]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        scope: { type: string }
        script_system_suffix: { type: string }
        image_prompt_prefix: { type: string }
        image_prompt_suffix: { type: string }
        image_model: { type: string }
        style_reference: { type: string }
        fps: { type: integer }
        width: { type: integer }
        height: { type: integer }
        codec: { type: string }
        panel_duration_ms: { type: integer }
        transition: { type: string }
...
    ProjectBody:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        defaults:
          description: Default Studio overrides hydrated when this project is picked.
          type: object
          additionalProperties: true
        archived: { type: boolean }
    ProjectView:
      type: object
      required: [id, name, description, defaults, archived, created_at, updated_at]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        defaults:
          type: object
          additionalProperties: true
        archived: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CharacterBody:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        traits:
          type: object
          additionalProperties: true
        ref_asset_ids:
          type: array
          items: { type: string }
    CharacterView:
      type: object
      required: [id, project_id, name, description, traits, ref_asset_ids, created_at, updated_at]
      properties:
        id: { type: string }
        project_id: { type: string }
        name: { type: string }
        description: { type: string }
        traits:
          type: object
          additionalProperties: true
        ref_asset_ids:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    EnvironmentBody:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        traits:
          type: object
          additionalProperties: true
        ref_asset_ids:
          type: array
          items: { type: string }
    EnvironmentView:
      type: object
      required: [id, project_id, name, description, traits, ref_asset_ids, created_at, updated_at]
      properties:
        id: { type: string }
        project_id: { type: string }
        name: { type: string }
        description: { type: string }
        traits:
          type: object
          additionalProperties: true
        ref_asset_ids:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SocialAccountBody:
      type: object
      required: [platform]
      properties:
        platform: { type: string, description: youtube_selenium | twitter_selenium | ... }
        label: { type: string }
        firefox_profile_path: { type: string }
        extra:
          type: object
          additionalProperties: true
    SocialAccountView:
      type: object
      required: [id, project_id, platform, label, firefox_profile_path, extra, created_at, updated_at]
      properties:
        id: { type: string }
        project_id: { type: string }
        platform: { type: string }
        label: { type: string }
        firefox_profile_path: { type: string }
        extra:
          type: object
          additionalProperties: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PlotBeatView:
      type: object
      required: [name, description, order]
      properties:
        name: { type: string }
        description: { type: string }
        order: { type: integer }
    PlotView:
      type: object
      required: [id, project_id, name, premise, beats, created_at, updated_at]
      properties:
        id: { type: string }
        project_id: { type: string }
        name: { type: string }
        premise: { type: string }
        beats:
          type: array
          items: { $ref: '#/components/schemas/PlotBeatView' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PlotBody:
      type: object
      properties:
        name: { type: string }
        premise: { type: string }
        beats:
          type: array
          items: { $ref: '#/components/schemas/PlotBeatView' }
    ProjectDetailView:
      type: object
      required: [project, characters, environments, social_accounts]
      properties:
        project: { $ref: '#/components/schemas/ProjectView' }
        characters:
          type: array
          items: { $ref: '#/components/schemas/CharacterView' }
        environments:
          type: array
          items: { $ref: '#/components/schemas/EnvironmentView' }
        social_accounts:
          type: array
          items: { $ref: '#/components/schemas/SocialAccountView' }
        plot:
          allOf:
            - { $ref: '#/components/schemas/PlotView' }
          nullable: true